+++ packages/react-query-devtools/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/react-query-devtools
 
+## 5.99.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.99.2
+  - @tanstack/react-query@5.99.2
+
+## 5.99.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.99.1
+  - @tanstack/react-query@5.99.1
+
 ## 5.99.0
 
 ### Patch Changes